pyproject.toml
[project]
name = "leverte_ia_tjenester"
version = "0.1.0"
description = "Datafortelling basert på data fra Fia"
authors = [{name="NAV-IT"}]
readme = "README.md"
requires-python = "==3.13.7"
dependencies = [
    "db-dtypes==1.4.3",
    "google-auth==2.40.3",
    "google-cloud-bigquery==3.35.1",
    "ipython==9.4.0",
    "jupyter-client==8.6.3",
    "jupyter-server==2.16.0",
    "logging==0.4.9.6",
    "notebook==7.4.5",
    "numpy==2.3.2",
    "pandas==2.3.1",
    "plotly==6.3.0",
    "requests==2.32.5",
    "tabulate==0.9.0"
]

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[project.urls]
Repository = "https://github.com/navikt/fia-datafortelling.git"

##### Pytest options
[tool.pytest.ini_options]
addopts = "-ra -q" # reduserer output så lenge alle tester består
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "nbstripout==0.8.1",
    "pre-commit==4.3.0",
    "pytest==8.4.1",
]
